Sorting bugs out in function stateChange
Current line 21 in function stateChange is
newArray = flipRandomSite(spin_array, size).
This one line is somehow setting other variables to be equal to it in this one line and I'm unsure how it is doing that. I commented out this line and I got my magnetization to remain constant, which since the lattice isn't changing that makes sense. Once I put in this one line again, my magnetization goes crazy. I started to print out when a change to my system was accepted and unaccepted, and I found that even when changes were not accepted, the system changed regardless.